Introduction of Window Switching/Dropdown Box work
Now that the button logic works and is easily created (by the grouping using qtobject), a new component called GeneralWindow is here such that when it is called, it will only appear when its id is called. (i.e "General" is made when any button in the general group is clicked). 3 more windows will be created for runs, startup and themes. You can also insert items that will also appear disappear.

The dropdown box works (i.e you can change the box) but im currently having issues with hierarchy (i.e the selection of the dropdown is appearing below another dropdown.